requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Flask-APScheduler==1.13.1
line-bot-sdk==3.9.0
requests==2.31.0
geopy==2.4.1 